separate home navgraph from main navgraph
