+++ presentation/home/src/main/java/com/giraffe/home/navigation/main/HomeRoute.kt
@@ -1,0 +1,17 @@
+package com.giraffe.home.navigation.main
+
+import com.giraffe.designsystem.composable.navbar.BottomTab
+import com.giraffe.designsystem.composable.navbar.Route
+import kotlinx.serialization.Serializable
+
+
+@Serializable
+data object HomeRoute : Route("home")
+
+class HomeTab(
+    override val labelRes: Int,
+    override val iconRes: Int
+) : BottomTab<HomeRoute> {
+    override val route = HomeRoute
+}
+
