+++ presentation/home/src/main/java/com/giraffe/home/screen/home/HomeRoute.kt
@@ -7,14 +7,7 @@ import com.giraffe.designsystem.composable.navbar.Route
 import kotlinx.serialization.Serializable
 
 @Serializable
-data object HomeRoute : Route("home")
-
-class HomeTab(
-    override val labelRes: Int,
-    override val iconRes: Int
-) : BottomTab<HomeRoute> {
-    override val route = HomeRoute
-}
+data object HomeRoute
 
 
 fun NavGraphBuilder.homeRoute(
